fix: throughput as f64

+++ crates/networking/p2p/sync/metrics.rs
@@ -12,7 +12,7 @@ struct ExecutionCycle {
     finished_at_block_hash: H256,
     executed_blocks_count: u32,
     add_block_time: u64,
-    throughput: u64,
+    throughput: f64,
 }
 
 impl Default for ExecutionCycle {
@@ -26,7 +26,7 @@ impl Default for ExecutionCycle {
             finished_at_block_hash: H256::default(),
             executed_blocks_count: 0,
             add_block_time: 0,
-            throughput: 0,
+            throughput: 0.0,
         }
     }
 }
@@ -57,7 +57,7 @@ impl Monitor {
         block_num: u64,
         block_hash: H256,
         add_block_time: u64,
-        throughput: u64,
+        throughput: f64,
     ) {
         self.current_cycle.executed_blocks_count += executed_blocks;
         self.current_cycle.add_block_time += add_block_time;
@@ -98,7 +98,7 @@ impl Monitor {
         let add_block_time_ratio = elapsed as f64 / self.current_cycle.add_block_time as f64;
         let blocks_per_second = self.current_cycle.executed_blocks_count as f64 / elapsed as f64;
         let throughput =
-            self.current_cycle.throughput as f64 / self.current_cycle.executed_blocks_count as f64;
+            self.current_cycle.throughput / self.current_cycle.executed_blocks_count as f64;
 
         tracing::info!(
             "[SYNCING PERF] Last {} blocks performance:\n\
@@ -158,7 +158,7 @@ impl SyncMetrics {
         last_block_number: u64,
         last_block_hash: H256,
         add_block_time: u64,
-        throughput: u64,
+        throughput: f64,
     ) {
         for monitor in &mut self.monitors {
             monitor.log_cycle(
